lambda: Integration test refactor
Split out separate S10 and Ponds Forge integration tests.

Use chokidar for file watching as it allows us to ignore the integration
test code when watching the unit tests.

Ponds Forge has changed the web site for lane swimming, so a new v2 API
will need to be written for it.